+++ core/src/uix/core.clj
@@ -15,12 +15,6 @@
 
 (def ^:private goog-debug (with-meta 'goog.DEBUG {:tag 'boolean}))
 
-(def ^:private optimizations-enabled? (volatile! true))
-
-(defmacro set-optimizations-enabled! [enabled?]
-  (vreset! optimizations-enabled? enabled?)
-  nil)
-
 (defn- no-args-component [sym var-sym body]
   `(defn ~sym []
      (let [f# (core/fn [] ~@body)]
@@ -28,10 +22,12 @@
          (binding [*current-component* ~var-sym] (f#))
          (f#)))))
 
+(def props-assert-fn (atom (core/fn [& args] true)))
+
 (defn- with-props-cond [props-cond props-sym]
   (when props-cond
     `{:pre ~(mapv (core/fn [spec]
-                    `(preo.core/arg! ~spec ~props-sym))
+                    `(@props-assert-fn ~spec ~props-sym))
                   props-cond)}))
 
 (defn- register-spec! [props-cond ns sym]
@@ -117,10 +113,10 @@
             var-sym (-> (str (-> &env :ns :name) "/" fname) symbol (with-meta {:tag 'js}))
             memo-var-sym (-> (str (-> &env :ns :name) "/" memo-fname) symbol (with-meta {:tag 'js}))
             [hoisted body] (-> (uix.dev/with-fast-refresh memo-var-sym fdecl)
-                               (aot/rewrite-forms :hoist? @optimizations-enabled? :fname fname :force? (:test/inline (meta sym))))]
+                               (aot/rewrite-forms :hoist? true :fname fname :force? (:test/inline (meta sym))))]
         (register-spec! props-cond ns sym)
         `(do
-           ~@(aot/inline-elements hoisted &env @optimizations-enabled? (:test/inline (meta sym)))
+           ~@(aot/inline-elements hoisted &env true (:test/inline (meta sym)))
            ~(if (empty? args)
               (no-args-component memo-fname memo-var-sym body)
               (with-args-component memo-fname memo-var-sym args body props-cond))