+++ core/src/uix/preload.cljs
@@ -1,5 +1,8 @@
 (ns uix.preload
-  (:require [uix.dev]))
+  (:require [uix.dev]
+            [uix.validate]))
+
+(uix.validate/set-props-assertion-enabled! true)
 
 (uix.dev/init-fast-refresh!)